requirements.txt
# Основные зависимости для RAG системы
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-community>=0.0.20
langchain-qdrant>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
qdrant-client>=1.7.0

# FastAPI и веб-сервер
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.12.0
httpx>=0.25.0

# Телеграм бот
python-telegram-bot>=20.0

# Обработка YAML
pyyaml>=6.0

# Embeddings и модели
sentence-transformers>=2.2.0
torch>=2.0.0

# Дополнительные зависимости для работы с документами (если нужно)
trafilatura>=1.6.0
markdownify>=0.11.6
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.0.0

python-dotenv>=1.2.1
